refactor: centralize auth components
