+++ soho/src/routes/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../../../src/contexts/AuthContext';
 
 interface ProtectedRouteProps {
   children: React.ReactElement;
