+++ soho/src/routes/Routes.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter, Navigate } from 'react-router-dom';
 import MainLayout from '../components/MainLayout.tsx';
-import AuthPage from '../pages/AuthPage.tsx';
+import AuthPage from '../../../src/pages/AuthPage.tsx';
 import Dashboard from '../pages/Dashboard.tsx';
 import ProtectedRoute from '../routes/ProtectedRoute.tsx';
